add ViewersList component
