+++ apps/web/src/components/ParticipantWindow/ParticipantWindow.tsx
@@ -1,12 +1,9 @@
 import React, { useCallback } from 'react';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import clsx from 'clsx';
-import { Typography } from '@material-ui/core';
 import { useAppState } from '../../state';
 import ParticipantWindowHeader from './ParticipantWindowHeader/ParticipantWindowHeader';
-import { useRaisedHandsMap } from '../../hooks/useRaisedHandsMap/useRaisedHandsMap';
-import { useViewersMap } from '../../hooks/useViewersMap/useViewersMap';
-import { RaisedHand } from './RaisedHand/RaisedHand';
+import ViewersList from '../ViewersList/ViewersList';
 import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
 import { sendSpeakerInvite } from '../../state/api/api';
 import { useEnqueueSnackbar } from '../../hooks/useSnackbar/useSnackbar';
@@ -31,18 +28,12 @@ const useStyles = makeStyles((theme: Theme) =>
     hide: {
       display: 'none',
     },
-    header: {
-      fontWeight: 'bold',
-      padding: '0 1em',
-    },
   })
 );
 
 export default function ParticipantWindow() {
   const classes = useStyles();
   const { appState } = useAppState();
-  const raisedHands = useRaisedHandsMap();
-  const viewers = useViewersMap();
   const { room } = useVideoContext();
   const enqueueSnackbar = useEnqueueSnackbar();
 
@@ -58,24 +49,10 @@ export default function ParticipantWindow() {
     [room, enqueueSnackbar]
   );
 
-  const viewersWithoutRaisedHands = viewers.filter(viewer => !raisedHands.includes(viewer));
-
   return (
     <aside className={clsx(classes.participantWindowContainer, { [classes.hide]: !appState.isParticipantWindowOpen })}>
       <ParticipantWindowHeader />
-      <div className={classes.header}>{`Viewers (${viewersWithoutRaisedHands.length + raisedHands.length})`}</div>
-
-      {raisedHands.map(raisedHand => (
-        <RaisedHand key={raisedHand} name={raisedHand} handleInvite={handleInvite} />
-      ))}
-
-      <div style={{ padding: '0.4em 1em' }}>
-        {viewersWithoutRaisedHands.map(viewer => (
-          <>
-            <Typography variant="body1">{viewer}</Typography>
-          </>
-        ))}
-      </div>
+      <ViewersList handleInvite={handleInvite} />
     </aside>
   );
 }
